+++ react/components/Spinner/index.js
@@ -1,14 +1,15 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import config from 'vtex-tachyons/config.json'
 import { baseClassname } from '../icon/utils'
 
 const radius = 40
 const circ = 2 * radius * Math.PI
 
-const Spinner = ({ secondary, size, block }) => (
+const Spinner = ({ color, size, block }) => (
   <svg
-    className={`${baseClassname('spinner')} ${block ? 'db' : ''}`}
+    className={
+      `${baseClassname('spinner')} ${!color ? 'c-action-primary' : ''} ${block ? 'db' : ''}`
+    }
     xmlns="http://www.w3.org/2000/svg"
     viewBox="0 0 100 100"
     preserveAspectRatio="xMidYMid"
@@ -20,7 +21,7 @@ const Spinner = ({ secondary, size, block }) => (
       cy="50"
       fill="none"
       r={radius}
-      stroke={secondary ? config.colors.white : config.colors.blue}
+      stroke={color || 'currentColor'}
       strokeWidth="10"
       strokeDasharray={`0 0 2 ${circ}`}
       strokeLinecap="round"
@@ -31,9 +32,11 @@ const Spinner = ({ secondary, size, block }) => (
         attributeName="stroke-dasharray"
         dur="1.5s"
         repeatCount="indefinite"
-        values={`0 0 2 ${circ};
+        values={
+          `0 0 2 ${circ};
           0 0 ${circ * 0.75} ${circ};
-          0 ${circ - 2} ${circ * 0.75} ${circ}`}
+          0 ${circ - 2} ${circ * 0.75} ${circ}`
+        }
       />
       <animateTransform
         attributeName="transform"
@@ -50,13 +53,12 @@ const Spinner = ({ secondary, size, block }) => (
 )
 
 Spinner.propTypes = {
-  secondary: PropTypes.bool,
   block: PropTypes.bool,
+  color: PropTypes.string,
   size: PropTypes.number,
 }
 
 Spinner.defaultProps = {
-  secondary: false,
   block: false,
   size: 40,
 }