+++ react/components/icon/Help/index.js
@@ -1,6 +1,5 @@
 import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
-import config from 'vtex-tachyons/config.json'
 import { calcIconSize, baseClassname } from '../utils'
 
 const iconBase = {
@@ -59,7 +58,7 @@ class Help extends PureComponent {
 }
 
 Help.defaultProps = {
-  color: config.colors['serious-black'],
+  color: 'currentColor',
   size: 16,
   solid: false,
   block: false,